add scss-lint
may not agree with me on this one, but if the whole point of a styleguide is to have consistency and predictability then linters for the visual language should be in it! 

This is not DIRECTLY related to the actual construction of a styleguide, but in my experience it is absolutely crucial to have a linter if you are thinking about making your own!

Cheers!